ComputerVision/requirements.txt
astroid==2.4.1
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongo==0.2.0
Pillow==7.1.2
pylint==2.5.2
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
